+++ module/patch-core.js
@@ -94,16 +94,6 @@ export async function PatchCore() {
     ]);
   };
 
-  // Workaround for unlinked token in first initiative on reload problem. No core issue number at the moment.
-  // if (Actor.config.collection && Object.keys(Actor.collection.tokens).length > 0) {
-  // Object.keys(Actor.collection.tokens).forEach((tokenId) => {
-  // let actor = Actor.collection.tokens[tokenId];
-  // for (let m of ["update", "createEmbeddedEntity", "updateEmbeddedEntity", "deleteEmbeddedEntity"]) {
-  // actor[m] = ActorTokenHelpers.prototype[m].bind(actor);
-  // }
-  // });
-  // }
-
   // Add combat tracker context menu options
   {
     const origFunc = CombatTracker.prototype._getEntryContextOptions;