+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added item script and macro calls to their advanced tabs, for users familiar with JavaScript
 - Sped up the loading of compendium browsers
 - Reduces memory usage of compendium browsers (at least client side, probably also partially server side)
+- Separated optional low-light vision multipliers on token configurations into dim and bright
 
 ## 0.78.8 (hotfix)